refactor: removed MUI, refactored File-structure, Cards are poor components now
