+++ react.components/package.json
@@ -3,11 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.8.1",
     "@fontsource/roboto": "^4.5.3",
-    "@mui/icons-material": "^5.5.1",
-    "@mui/material": "^5.5.1",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
@@ -29,7 +25,6 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
-    "ts-jest": "^27.1.4",
     "typescript": "^4.6.2",
     "web-vitals": "^2.1.4"
   },
