+++ react.components/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './styles/App.css';
 import { BrowserRouter } from 'react-router-dom';
-import Navbar from './components/UI/Navbar/Navbar';
+import Navbar from './components/Navbar/Navbar';
 import AppRouter from './components/AppRouter';
 
 function App() {
