+++ react.components/src/components/Card/Card.module.css
@@ -1,0 +1,66 @@
+.card {
+    display: grid;
+    align-items: center;
+    flex-wrap: wrap;
+    width: 300px;
+    margin: 1em;
+    border-radius: .5em;
+    box-shadow: 0 0 1em #0000003d;
+    height: 450px;
+    transition: all 0.2s linear;
+}
+
+.card:hover {
+    box-shadow: 0 0 3em #0000003d;
+    height: 450px;
+}
+
+.cardImg > img {
+    vertical-align: middle;
+    height: 300px;
+    width: 100%;
+    border-radius: .5em .5em 0 0;
+}
+
+.cardMataInfo {
+    padding: 15px;
+    height: 70px;
+    overflow: hidden;
+}
+
+.cardMataInfo::-webkit-scrollbar {
+    width: 0;
+}
+
+.cardMataInfo > div:first-child {
+    font-size: 1.5em;
+    margin-bottom: .2em;
+}
+
+.cardMataInfo > div:nth-child(2),
+.cardMataInfo > div:nth-child(3)
+{
+    display: flex;
+    font-size: .8em;
+}
+
+.cardMataInfo > div:nth-child(2) > div:first-child,
+.cardMataInfo > div:nth-child(3) > div:first-child {
+    margin-right: .5em;
+    color: #979797;
+}
+
+.cardControls {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: nowrap;
+    align-items: center;
+    align-content: center;
+    justify-content: flex-start;
+
+    margin: 0 10px 10px 10px;
+}
+
+.cardControls > button:nth-child(3) {
+    margin-left: auto;
+}
