+++ react.components/src/components/CardSet/CardSet.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Card from './Card/Card';
+import Card from '../Card/Card';
 import ItemStyles from './CardSet.module.css';
-import { ICharacterRowInfo, ICardSetProps, ICardSetState } from '../../../types/types';
+import { ICharacterRowInfo, ICardSetProps, ICardSetState } from '../../types/interfaces';
 import axios from 'axios';
 
 class CardSet extends React.Component<
