+++ react.components/src/components/SearchBar/SearchBar.module.css
@@ -9,7 +9,7 @@
     background-color: white;
     background-size: 25px;
     background-position: 18px 14px;
-    background-image: url('../../../images/search-icon.svg');
+    background-image: url('../../asserts/svg/search-icon.svg');
     background-repeat: no-repeat;
     padding-left: 2.5em;
     font-size: 1.4em;
