+++ react.components/src/components/UI/ButtonCustom/ButtonCustom.module.css
@@ -1,0 +1,17 @@
+.btnCustom {
+    background-color: #ffa5009c;
+    color: white;
+    font-size: 1em;
+    border: none;
+    border-radius: 0.3em;
+    padding: 0.3em;
+    cursor: pointer;
+    margin: 0.5em;
+    box-shadow: 0.1em 0.1em 0.5em rgba(0, 0, 0, 0.2);
+    transition: all 0.2s linear;
+}
+
+.btnCustom:hover {
+    background-color: rgba(220, 139, 0, 0.71);
+    box-shadow: 0.1em 0.1em 0.5em rgba(0, 0, 0, 0.5);
+}
