+++ react.components/src/components/UI/ShareButton/ShareButton.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { IShareButtonProps } from '../../../types/interfaces';
+import ItemStyles from './ShareButton.module.css';
+import shareIcon from '../../../asserts/svg/share-icon.svg';
+
+class ShareButton extends React.Component<IShareButtonProps> {
+  constructor(props: IShareButtonProps) {
+    super(props);
+  }
+  render() {
+    return (
+      <button className={ItemStyles.shrBtn}>{/*<img src={shareIcon} alt="sher-btn" />*/}</button>
+    );
+  }
+}
+
+export default ShareButton;
