+++ react.components/src/tests/CardSet.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import CardSet from '../components/UI/CardSet/CardSet';
+import CardSet from '../components/CardSet/CardSet';
 import axios, { AxiosResponse } from 'axios';
-import MockDataSet from '../DataFromDB/MockDataSet';
-import { ICharacterRowInfo } from '../types/types';
+import MockDataSet from '../data/cardsData';
+import { ICharacterRowInfo } from '../types/interfaces';
 
 jest.mock('axios');
 const mockedAxios = axios as jest.Mocked<typeof axios>;
